requirements.txt
azure-mgmt-resource==1.1.0
azure-mgmt-network==1.4.0
azure-mgmt-compute==2.1.0
azure-mgmt-storage==1.2.1
requests[security]==2.18.4
haikunator
